Replace testId with data-automation attribute
Forgot testId attribute is from React Native, and the browser (Chrome) 
dumps a warning in the console for using an invalid attribute name.

In a way, it feels like an aria attribute might be a good fit for this, 
in a way it is making the markup accessible to another audience, a test system, even though it’s not human.

But without knowing much about screen readers, I guess there is a risk
inventing, or overlapping an existing aria- attribute, with a test 
value might just generate noise for users on devices dependent on them.