Add adjustments reducer based on derived card state
This commits introduces adjustments, which are basically just a record 
of a price reduction applied to an item based on an eligible and 
applicable promotion, as a first class object in the cart.

The implementation is based on the key assumption, that at any time,
there can only be one promotion for a given user & product combination.

For brevity, some test remain to be implemented.

I’ve also introduced get() functions to the cart reducer, so we don’t
need to expose the cart’s state structure through the app.